Production

Approved-by: Arjun Prajapati <arjun.prajapati@paytmmall.com>
Approved-by: Gaurav Das <gaurav.das@paytm.com>
Approved-by: Karan Verma <karan2.verma@paytmmall.com>

+++ digital-education-explore/src/main/java/com/paytm/digital/education/explore/es/model/InstituteSearch.java
@@ -67,7 +67,7 @@ public class InstituteSearch {
     private float maxRating;
 
     @JsonProperty("max_rank")
-    private int maxRank;
+    private long maxRank;
 
     @JsonProperty("courses")
     private List<NestedCourseSearch> courses;
